Move container / result objects out to separate header.
